Fixes zipped test case upload and download
Changed the structure of imported and exported test cases, and how it is
specified. New specification.yaml details how each test set contains the
points and the test cases it includes.

Test sets and cases must now have unique names, and these names are used
in the YAML specification to describe the graph of test sets/cases.

Importer and Exporter classes allow easier extension of different
import/export formats in the future.