Only install 'bundler' system-wide if 'gem' was installed system-wide
Only use sudo with 'gem install' if necessary. If 'gem' is in a user
directory (e.g. with RVM/rbenv) sudo won't find it.

Fix the path to bundler. The directory in a gem containing executables
doesn't have to be called 'bin'[1], and for bundler it's currently
'exe'. But the executable in that directory shouldn't be used directly
anyway, instead use the RubyGems binstub[2] in Gem.bindir[3] (usually
the directory containing the 'ruby' executable).

[1] https://guides.rubygems.org/specification-reference/#bindir
[2] https://github.com/rbenv/rbenv/wiki/Understanding-binstubs
[3] https://ruby-doc.org/stdlib/libdoc/rubygems/rdoc/Gem.html#method-c-bindir