Detect bundle errors in install scripts that use pipes '|'
The return status of a pipeline such as 'bundle install | sed ...' is
the exit status of the last (rightmost) command by default, in this case
'sed'. If the bundle command fails the error will be ignored.

The 'pipefail' option of bash fixes this. When enabled, the pipeline's
return status is the the exit status of the last command that failed.