Disable async loading of JavaScript in development mode
In development mode,

    <%= javascript_include_tag "application", :async => true %>

expands to several <script> elements and each has the "async" attribute.
They can load out of order which causes errors such as "jQuery is not
defined" in the browser's web console.

The async attribute should not be set in development mode. It's fine in
production because the files are concatenated into a single script.

For more details, see
https://rhardih.io/2015/03/rails-4-gotcha-asynchronous-loading-of-javascript-in-development-mode/