Fix background colour of test case output in submission page
The test case inspection element (containing the actual and expected
output, shown on mouse hover) is a sibling of the test case element,
so its background colour is inherited from the parent test set element
rather than the test case.

This becomes apparent when a test set contains a mix of correct and
wrong answers. The background of the inspection element will be red
even for the correct test cases, because the parent test set is red.

Fix this by adding the status class to the inspection element.