Fix menus on touch devices
Menus display on hover, and some also have a click action (e.g.
username in the top right corner is both a menu and a link to the
profile page; community > users is both a submenu and a link).

On some touch devices (e.g. Chrome on Android) this is unusable:
tapping displays the menu but then also performs the click action. The
first tap should display the menu and a second tap should perform the
click action.

Superfish (the menu plugin used) already supports this behaviour, but
to work on more devices it needs the hoverIntent plugin. This commit
enables that plugin.

In the case of the username there is an additional problem: the <a>
element is inside a <span> element, and Superfish doesn't support
that. The <span> element was introduced in commit 0e2fb4e ("refactor
submissions to problems", 2012-08-29) and does not appear to be
essential, so this commit removes the <span> element. Its ID is used
in a test, so move the ID to the <a> element.

Removing the <span> element from the username has the side effect of
adding an arrow symbol (like the other menus).