Fix double destroy in submission spec
The spec creates a problem and submissions, then destroys the problem
and the submissions. However, the `submissions` association on
`Problem` uses `dependent: :destroy`, so destroying the problem
automatically destroys the submissions. When the submissions are then
destroyed explicitly it causes an error:

An error occurred in an after(:all) hook.
  NoMethodError: undefined method `recalculate_tests_and_save!' for nil:NilClass
  occurred at ./app/models/submission.rb:50:in `block in <class:Submission>'

Fix this by destroying the objects in reverse order of creation, as
done in contest_score_spec.rb. Another option is to not explicitly
destroy the submissions, but that breaks the create/destroy symmetry.