Use 'if cond; then cmd; fi' instead of 'cond && cmd'
I know the scripts use '&&' a lot to conditionally run commands, but
it's actually an anti-pattern. With '&&', when 'cond' fails the entire
command fails, which is usually not the intention.

This can become significant when it's the final command in the script,
in which case it sets the exit code of the script.