Improve regex for converting .NET Core SDK version to .NET Core Runtime version
From to the documentation[1], the last component of the SDK version is
of the form (minor_version * 100 + patch_number) and minor_version can
be more than one digit, though it's unlikely to occur. (Also, the
other components aren't guaranteed to be single digits, and '.' should
be escaped.)

There is still an issue because the minor_version in the SDK version
does not necessarily correspond to the last component of the Runtime
version.

[1] https://docs.microsoft.com/en-us/dotnet/core/versions/