Add 'Edit' button to current user's profile (only for non-admins)
The button points to /accounts/settings/edit, like "edit profile" in
the user menu, to make it easier for users to edit their own profile.

Admins already have an 'Edit' button on all user profiles that points
to /user/:id/edit, and that stays the same (including on their own
profile) for consistentcy with the 'Edit' links in /users. Admins can
use "edit profile" from the user menu to get to
/accounts/settings/edit.