Require school to be blank if outside NZ
Users outside New Zealand are considered to be unofficial contestants
for NZIC. However, students outside New Zealand were allowed to select
a school. Requiring a blank school allows us to remove the country
check when determining whether a user is an official contestant.

A user is considered to be official by the presence of a school and
school year from their contest relation (not their profile), since a
user may change schools/countries after the contest.