Change visibility of problem statement attachment to 'public' in importer (#112)
The problem series import worker previously set the visibility of the
file attachment containing the imported statement to 'protected'.

That doesn't make sense, because 'protected' means users can view the
file if they are *not* currently competing (comment: [1], logic: [2]).

As discussed in #106 (specifically comment [3] by @Holmes98), the
purpose of 'protected' is probably to prevent access to file
attachments of other problems/groups during closed-book contests. It
does not make sense to use it for problem statements, because users
need to be able to view them during the contest.

This commit changes the visibility to 'public'. It is safe (in the
sense that it won't make the problem statements visible before the
user starts the contest) because the controller also checks that the
containing problem is accessible.[4]

[1] app/models/filelink.rb:17
[2] app/policies/filelink_policy.rb:26
[3] https://github.com/NZOI/nztrain/issues/106#issuecomment-617556022
[4] app/controllers/filelinks/roots_controller.rb:53