Remove invalid <p> element around problems table in contest page
Previously the markup for the "problems" tab on the contest page was

    <p>
    <table>
    ...
    </table>
    </p>

This is invalid because <table> is a block-level element and can't
appear inside a <p> element. It's actually parsed as

    <p></p>
    <table>
    ...
    </table>
    <p></p>

The effect is that it adds spacing before and after the problems
table. I don't think it's intentional (the other tables don't have
spacing at the end), so I'm removing the <p> tags.