Write confirmation emails to a file instead of mailing them, as a temporary workaround
We use a Gmail account to send confirmation emails for new accounts,
but since May 30, 2022 Google has disabled app authentication using
username and password[1] and this causes account registration to fail.

As a temporary workaround, write the emails to a file instead of
sending them, so staff can copy the email and send it manually.

The emails will be written to <app-root>/tmp/mails/<email-address> (if
multiple emails are sent to the same account they are concatenated).

Update the confirmation instructions to ask users to contact staff so
we can manually send them the confirmation link.

[1]: https://support.google.com/accounts/answer/6010255