Deliver emails to a Discord channel using a webhook
This is nicer than writing the emails as files because we can receive
notifications from Discord (so students shouldn't need to send us a
request) and because staff will be able to view and forward the emails
without requiring SSH access to the server.

The webhook URL can be created in Discord by opening the Server
Settings then the Integrations tab:
https://support.discord.com/hc/en-us/articles/228383668-Intro-to-Webhooks

This URL needs to be set in the web app using admin > settings > New
*before* this commit is deployed. The key is "system/mailer/webhook"
and the value is the webhook URL.

This commit also updates the messages in the interface to only ask
students to contact staff if they don't get the email within a day.

Here are some links to relevant documentation.

ActionMailer delivery_method:
https://api.rubyonrails.org/v4.0.13/classes/ActionMailer/Base.html
https://github.com/mikel/mail/tree/2-5-stable#sending-an-email
https://www.rubydoc.info/gems/mail/2.4.0/Mail/FileDelivery#deliver!-instance_method
https://github.com/mikel/mail/blob/2-5-stable/lib/mail/network/delivery_methods/file_delivery.rb
https://www.rubydoc.info/gems/mail/2.4.0/Mail/Message

Ruby HTTP client:
https://ruby-doc.org/stdlib-2.3.8/libdoc/net/http/rdoc/Net/HTTP.html#class-Net::HTTP-label-How+to+use+Net-3A-3AHTTP
https://ruby-doc.org/stdlib-2.3.8/libdoc/net/http/rdoc/Net/HTTP.html#class-Net::HTTP-label-Setting+Headers

Discord webhooks:
https://support.discord.com/hc/en-us/articles/228383668-Intro-to-Webhooks
https://discord.com/developers/docs/resources/webhook#execute-webhook
https://discord.com/developers/docs/reference#message-formatting
https://github.com/discord/discord-api-docs/ issue 1286 (the exact message sanitization rules are intentionally undocumented)