Give group members access to group scoreboard (#146)
The group scoreboard page shows the scores of members on all the
problems in the group. Previously it was only visible to admins. This
commit gives group members access to this page because we hope it
would be motivating for students to see the progress of their peers.

The group scoreboard has a subtle feature where a completely blank
cell indicates that the user has never viewed the problem. This commit
disables this feature for non-admins, to give users some privacy from
other group members.

This commit also hides the submission counts when the group scoreboard
is viewed by a non-admin, again to give some privacy. (The existing
contest scoreboards show the submission counts, but the privacy
expectations may be different when practising as opposed to
competing.)

Co-authored-by: Tom Levy <tomlevy93@gmail.com>