Fix line number issue when displaying submission source etc. (#168)
The recent upgrade to pygments.rb 1.1.2 (#164) caused a small issue
with the line numbers on syntax-highlighted code blocks (including
test case inputs and outputs). It caused the numbers "1" and "2" to
both appear next to the first line on top of each other (see the PR
for a screenshot).

This happened because of a change in Pygments that intentionally
inserted an empty <span> element at the beginning to preserve leading
empty lines.[1][2]

The line numbers are added by our CSS, and the old selector treated
this dummy span as a line.

The spans that correspond to actual lines have an ID attribute of the
form "line-1", "line-2" etc. So this commit fixes the issue by
changing the selector to only match spans with an ID attribute
starting with "line-".

[1]: https://www.github.com/pygments/pygments/issues/818
[2]: https://www.github.com/pygments/pygments/issues/961