Don't redirect to sign in if the client is not a web browser
Previously, when a user who is not signed in requests a restricted
page, we'd redirect to the sign-in page. (In comparison, signed-in
users get a "403 Forbidden" page.)

This is great when the client is a web browser, because the user
probably wants to sign-in so they can view the restricted page.

But it is not appropriate for automated clients (e.g. wget), which
will probably follow the redirect and fetch the sign-in page without
indicating that there was an authentication failure.

Another problem is that if the client requests a non-HTML MIME type by
setting the HTTP "Accept" header to e.g. "application/xml", then when
following the redirect the response will confusingly be "406 Not
Acceptable" (because the sign-in page only supports HTML).

To fix these issues, try to detect whether the client is a web browser
by checking if the "Accept" header includes "text/html", and if the
client is not a web browser return "403 Forbidden" instead of
redirecting.