Specify the Ruby version in .ruby-version and Gemfile
The file .ruby-version is used by rbenv, RVM, asdf[1], and
setup-ruby[2] to choose the Ruby version.

Also specify the supported Ruby versions in the Gemfile for tools such
as Dependabot, which does not currently support .ruby-version and
would otherwise suggest incompatible updates.

The versions in the two files will have to be kept in sync.
Fortunately, Bundler reports an error if the version specifier in the
Gemfile is not satisfied, so it should be very obvious if they are out
of sync.

The install scripts should be updated to use .ruby-version instead of
a hard-coded version. This will be done separately.

Closes #170.

[1]: https://github.com/asdf-vm/asdf-ruby
[2]: https://github.com/ruby/setup-ruby (for GitHub Actions)