Remove MaxMind GeoIP database and related code (#108)
The MaxMind GeoLite2 database requires a (free) license key since 30
December 2019 due to issues related to privacy regulations. See [1]
and issue #88 for details.

The GeoIP database is only used to set the default country on the
registration page and in an obsolete database migration that guesses
the country of users who registered before the country field was added
to the registration form.

It's not worth the hassle of obtaining a license key and maintaining
the install script.

After this commit, New Zealand will become the default country on the
registration page for all users.

(Side note: The maxmind.bash script also has a small bug. It uses
add-apt-repository before installing software-properties-common.)

Closes #88.

[1] https://blog.maxmind.com/2019/12/18/significant-changes-to-accessing-and-using-geolite2-databases/