Allow redis config to use ERB tags `<%= %>`
This makes dockerisation much easier, because we can create a
redis.yml file (included in the image) that looks up a specific
environment variable, and then we can set that environment variable
when we run the image. This is much easier than generating redis.yml
at startup.

(Rails already allows config/database.yml to contain ERB tags, so we
are just following their lead.)

Co-authored-by: Tom Levy <tomlevy93@gmail.com>