Avoid a local jump error caused by trying to return from a block
Returning from a block isn't really supported - instead you're
supposed to use break / next to exit early from a block. In any rate,
this callback just needs to return true, which we can do via implicit
return.

Rails 4.1 treats a non-local jump in an ActiveRecord::Callback as an
error, so we have to fix this before we upgrade.

Note that Rails 5.1 ignores the callback's return value (the new
approach is an explicit `throw :abort`), so once we upgrade to Rails
5.1 the return value will become superfluous.

Co-authored-by: Tom Levy <tomlevy93@gmail.com>