Remove the AI Contest implementation (#177)
The idea of the AI Contest is to have two programs compete against
each other in a game. This code is currently unused and it would be
difficult to get it working properly.

We have a lot of other code to upgrade, so we are removing the AI
Contest for now. We might bring it back later after some more thought
about how exactly to maintain it (see #185 for a rough attempt).

Before deploying this commit, the AI Contest tables should be backed
up using `pg_dump` and the migrations should be brought down using the
following command (which drops the tables):

```
for i in $(cd db/migrate; ls -r *_ai_* 20130113010505_add_uniq_index_to_games.rb); do
    RAILS_ENV=production bundle exec rake db:migrate:down VERSION=$i || break
done
```

Co-authored-by: Tom Levy <tomlevy93@gmail.com>