Downgrade json to 1.8.6 for Rails 4.1 compatibility; add workaround for CVE-2020-10663
Rails' activesupport 4.1.* depends on json ~> 1.7 (i.e < 2.0) [1], so
in order to upgrade to Rails 4.1 we need to downgrade json from 2.6.3
to 1.8.6 (the highest < 2.0).

But json < 2.3 has a security vulnerability, CVE-2020-10663 [2]. So we
also add a dependency on the gem 'json_cve_2020_10663', which
monkey-patches the json gem to protect from this vulnerability.

(Verified that the CVE is patched using the code listed on [3].)

[1] https://rubygems.org/gems/activesupport/versions/4.1.16
[2] https://www.ruby-lang.org/en/news/2020/03/19/json-dos-cve-2020-10663/
[3] https://github.com/rails-lts/json_cve_2020_10663

Suggested-by: Jonathan Khoo <jkhoo98@gmail.com>