Don't indent output in install scripts (#186)
Indenting the output of verbose commands such as `rake db:migrate` is
nice because it visually separates their output from everything else,
but there are a couple of problems with the current approach:

1. It only indents stdout, so messages written to stderr are
   de-synchronized and can appear out of order, which is confusing.
   [Can be fixed by also indenting stderr: `... |& sed ...`]

2. Some commands (e.g. `bundle install`) use colour when their output
   is a terminal. Piping their output suppresses this behaviour, which
   is a shame. [Can't be fixed easily]

3. The complex filter for `bundle install` is broken:

   - It tries to remove lines of the form "Using <gem> (<version>)",
     but the regex is broken (Bundler's current output format doesn't
     contain parentheses) so these lines are still displayed.

   - It buffers the entire output in memory, so there is no indication
     of progress (the output appears all at once when Bundler exits).

   - The buffering behaviour also breaks the indentation. [Easy fix]

   [All can be fixed by switching to `sed -e '/^Using/d; s/^/   /g'`]

As noted inline, these issues are easy to fix (except for colour
output). But I don't think the indentation is very useful, so I'd
rather just drop it altogether.

Regarding the "Using" lines in output of `bundle install`: Bundler now
uses a different colour for the other lines, which helps. And I think
the "Using" lines can be useful (e.g. for debugging gem versionining
issues). And I'm slightly uncomfortable with the fact that the script
displays `$ bundle install` but then runs a different command with
modified output.