Add config.secret_key_base (#195)
secret_key_base was added in Rails 4 and replaces secret_token. It is
used by CookieStore to encrypt cookies (we are not currently using
CookieStore, but we will switch to it in #162).

The Rails 4 convention for setting secret_key_base is to use
config/secrets.yml and the environment variable SECRET_KEY_BASE (see
[1]). We don't use this approach, because secrets.yml will be
deprecated in favour of credentials.yml.enc in Rails 5.2 [2] and
because setting the environment variable is awkward.

Instead, we generate a random secret the first time the app runs and
store it in config/secret_key_base.#{env}.txt. We also support the
environment variable SECRET_KEY_BASE to match the Rails convention; we
don't currently use this environment variable, but it may be useful
(e.g. in situations such as build steps where we don't want to
generate the file, or in environments such as Docker where environment
variables are often more convenient than files).

We don't store the secret in the database (which is how the existing
secret_token is implemented) because it may be useful to start the app
without a database connection. Also, it's probably more secure to keep
the secret out of the database (and its backups).

This commit also adds a rule to .gitignore to make sure the secrets
aren't accidentally committed.

[1]: https://guides.rubyonrails.org/v4.1.8/upgrading_ruby_on_rails.html#config-secrets-yml
[2]: https://www.github.com/rails/rails/pull/30067