Remove config.secret_token (#195)
It has been replaced by secret_key_base, and doesn't appear to be used
anywhere because we are not using CookieStore yet.

Also rename the file because it is no longer related to secret_token.

The setting sessions/secret_token should eventually be deleted
manually using the admin web interface, but note that this setting is
still used in config/initializers/devise.rb at the moment.