Revert "Remove config.secret_token (#195)"
I was wrong, we *are* using CookieStore (just not for the session
store). We need to keep secret_token for now, to allow automatic
migration from the legacy signed-but-not-encrypted cookies to the new
store encrypted using secret_key_base.

The cookies are used for the "remember me" functionality for example.