Limit /users/online to staff
The "community" nav menu only shows the "online" link to staff,
however there is no corresponding authorization check to prevent
non-staff from accessing that page via a direct link. This commit adds
the check.

(Later on we'll add more restrictions the other "community" pages due
to privacy. We are starting with the "online" page because it's not
even visible in the nav menu, and because it's the only page that
shows the last login time -- which is on the more sensitive side.)

See: #192

Reported-by: Bruce C <brucecccccc@gmail.com>