Make /users etc. twice as fast by preloading roles
This is an example of the N+1 query problem. Fixing it makes the pages
render about twice as fast when there are many (thousands) of users.

The rest of the time is spent in the view, which is harder to optimise
(there are some low-hanging fruit such as caching the policies, but
they'll only get us so far; the only complete solution is to introduce
paging).

See: #192

Reported-by: Bruce C <brucecccccc@gmail.com> (and others)