Update to Rails 4.2 (#199)
Steps:
* Update Gemfile
* bundle update --conservative rails actionpack railties activerecord activemodel activesupport rack responders

Also add an explicit dependency on responders, because Rails 4.2 moved
the `respond_with` method out to a separate gem[1]. (We're currently
okay because devise depends on responders, but seeing how we're
explicitly using the `respond_with` method we should add an explicit
dependency.)

There are lots of warnings & deprecations, but nothing is _failing_ in
the test suite. The deprecation warnings will be fixed separately
(e.g. #219, #220).

Note that due to a change in Rack, `rails server` now listens on
`localhost` instead of `0.0.0.0` by default in development mode[2][3].
This prevents connections from other machines for security. The older
behaviour can be achieved using `rails server -b 0.0.0.0`. (Also, when
using `rails server` without `-b`, note that `localhost` may resolve
to an IPv6 address depending on the operating system, in which case
http://127.0.0.1:3000 won't work; instead, use http://localhost:3000.)

[1] https://guides.rubyonrails.org/4_2_release_notes.html#respond-with-class-level-respond-to
[2] https://guides.rubyonrails.org/4_2_release_notes.html#default-host-for-rails-server
[3] https://github.com/rack/rack/commit/28b014484a8ac0bbb388e7eaeeef159598ec64fc

Co-authored-by: Tom Levy <tomlevy93@gmail.com>