Setup capistrano for deployments
The current deployment process doesn't give us many options for
rollbacks, and could have issues if the server was restarted during the
deployment. Most rails apps that aren't containerised tend to be
deployed with Capistrano.

Capistrano lets us all the same user, and manage deployment access by
giving people access to the deploy user by adding their ssh public key.
We should be able to move towards that being a sudo-less enterprise, but
qless may make that difficult.

Update the puma config in prep for actually running in production, based
on the recommended unit file from the Puma repo. Add sd_notify to the
gemfile to so that puma can use systemd watchdog to make sure it's
running healthy and happy.

The current deployment config is only for a "staging" environment
running on a virtual machine at Ben's house. We can use this
indefinitely as "staging", then add a "production.rb" when we want to
use capistrano for deployments to production.