Change how we check if PostgreSQL server is installed (#237)
Previously we used `psql --version`, which actually checks the client
utility rather than the server. Switch to checking whether the
`postgres` user exists, since that's a simple and fairly reliable
test.

This commit drops the version check, because it's fairly tricky to
check the server version (and we don't have much control over the
version anyway because we are installing the package from the
archives). Some ideas that were discarded:

 - `postgres --version`: Fails because postgres is not in $PATH.

 - `pg_config --version`: This is supposed to print the server
   version, but it succeeds even if the server is not installed (e.g.
   when the libpq-dev package is installed).

 - `sudo -u postgres psql -t postgres -c 'SELECT version();'`:
   If the server is installed but not running, this fails. But in that
   case we probably shouldn't try to install the server.