Add env var DATABASE_INSTALL to skip PostgreSQL install (#237)
If the database is running on another host/container, then we don't
need to install the PostgreSQL server on the web server.

This is useful for CI, and mirrors the existing REDIS_INSTALL
environment variable.

We always need the PostgreSQL client utilities, so install them
separately.

If we skip installation of the database then we don't try to create
the database user, because that's unlikely to work with remote
databases.