Restore :assets group in Gemfile
Some of those gems (especially mini_racer) are very large, and it is
useful to be able to skip them. They were already in a commented-out
group; restore the group to allow skipping them (they will still be
installed by default).

Note that putting them in a group will mean they won't be
auto-required when config/application.rb calls
`Bundler.require(:default, Rails.env)`, but that should be fine.

Also remove the old comment "not required in production environments
by default" because it is unclear/inaccurate (we currently run the
asset precompilation on the production server), and add hints on how
to skip the group.