Send deprecation notices to the rails log
Prior to this change, messages were being send to the
ActiveSupport::Notifications system which doesn't appear to have any
active listeners, meaning these notices were just being sent into the
void. The notification was probably setup for a old integration (maybe
newrelic?) but nothing appears to be using it at the moment.