Don't use --skip-update with install.bash in CI
This option was inherited from .travis.yml and causes install.bash to
skip the following steps (which are also done during update):

 - bundle install
 - script/install/migrate.bash
 - script/install/whenever.bash
 - (if in production) script/install/assets.bash

The job currently works even without these steps, but it's fragile:
the `bundle install` step is only safe to skip because we pass
`bundler-cache: true` to ruby/setup-ruby; if we disable the cache then
ruby/setup-ruby won't run `bundle install` and the build will fail.

So don't use --skip-updates, in order to always run `bundle install`
and ensure the build will work even without `bundler-cache: true`.
(Running `bundle install` twice doesn't do any harm, it only takes
~0.27s the second time.)

Not using --skip-updates has the a additional effect of running
migrate.bash, which is probably a good thing because it exercises the
migration code. The migrations take ~8s, which is a little slow but
not a big deal. Note that the schema is dumped after migrating, so
db:test:load will use the dumped schema rather than the one from the
repository. We will add a check later on that reports an error if the
dumped schema has any differences.

Running whenever.bash is quick and doesn't do any harm.