Convert `t.references` to int + string columns for `limit: 255`
In a table definition, `t.references :mycol, polymorphic: true` is a
shorthand for defining two columns, `t.integer :mycol_id` and
`t.string :mycol_type` [1].

We are going to change the migrations to explicitly set `limit: 255`
on string columns (see the next commit for details). There is no way
to directly set `limit: 255` on the string column defined by
`t.references`, so we expand it out into the explicit integer and
string column definitions.

(This commit doesn't actually set `limit: 255`; that will be done in
in the next commit.)

[1] https://github.com/rails/rails/blob/v4.2.0/activerecord/lib/active_record/connection_adapters/abstract/schema_definitions.rb#L300-L319

Suggested-by: Ben Anderson <me@benanderson.nz>