Explicitly set `limit: 255` on string columns in migrations
There used to be a default limit of 255 characters on string columns,
but that was removed in Rails 4.2 [1][2].

Explicitly set the limit to 255 in the migrations so that they will
retain their old behaviour.

Note that we actually plan to remove the limit (and match the new
default), but that will be done separately.

Also note that this commit only changes the migrations -- the schema
also needs to explicitly set the limits, and that will be done in the
next commit.

This change was made using the following commands:

```
sed -i -E '/limit:/! { /_column/ s/:string/:string, limit: 255/; }' db/migrate/*
sed -i -E '/limit:/! { s/t\.string *:\w*$/&, limit: 255/; }' db/migrate/*
sed -i -E '/limit:/! { s/t\.string *:\w*,  */&limit: 255, /; }' db/migrate/*
```

Explanation:

All three commands ignore lines that already set a limit. The first
command is for lines such as `add_column :users, :email, :string`. The
second command is for lines such as `t.string :email`. The third
command is for lines such as `t.string :email,     null: false`
(taking care to preserve alignment).

[1] https://guides.rubyonrails.org/v4.2/4_2_release_notes.html#active-record-notable-changes
[2] https://www.github.com/rails/rails/pull/14579