Fix minor bug in /submissions using "My" for user 0
The title of /submissions/by_user/:by_user is "My Submissions" when
:by_user is the current user, and it's *supposed* to be "Submissions"
in all other cases. (And a similar thing with the messages "You have
no submissions." / "No submissions found.")

But there is a minor bug: When user 0 views /submissions (which
displays submissions from all users), the title is "My Submissions".
This is because `params[:by_user].to_i` evaluates to 0 when :by_user
missing, which is equal to `current_user.id` for the System user.

Fix this by explicitly checking whether :by_user is present.

(In fact, it's probably more consistent to only use the title "My
Submissions" when viewing /submissions/my, and use the title
"Submissions" for /submissions/by_user/:by_user even if :by_user is
the current user. That will be done in the next commit.)