Switch from `authorize Submission.new, :show?` to `:inspect?`
The difference is that :show? also returns true on submissions in the
policy scope (e.g. submissions to problems owned by the current user),
but that's irrelevant when giving it a dummy Submission. So it's a
little odd to use :show?.

(It could be argued that :show? is more correct. But then the logical
conclusion is that /submissions should be permitted for all users and
should show them all the submissions they can view. But that's not the
case.)