Remove hard-coded default SMTP user name and password
Passwords should never be hard-coded in the source code for security.
The default user name was somewhat useful for the official instance,
but it's cleaner to avoid it.

(Note: Now that the default value is blank, it's not clear that the
setting "system/mailer/username" is supposed to be an email address.
This will be addressed in a later commit that renames that setting to
"system/mailer/email".)