Set "from" email address using settings table instead of hard-coding
Previously the "from" address was hard-coded to nztrain@gmail.com and
could not be configured.

This commit uses the value of the "system/mailer/username" setting,
which is the user name for SMTP authentication. It will be the same as
the "from" email when using Gmail and in most other setups. (A later
commit will rename that setting to "system/mailer/email".)

* We provide a default email in the test environment, otherwise the
  specs fail.
* We don't provide a default email in the production environment,
  because the admin must configure email properly in order
  for mail delivery to work.
* We don't provide a default email in the development environment,
  because mail delivery will fail anyway (unless SMTP is configured
  properly). The admin will still be able to read the emails by
  looking in the development log.

In the future we may wish to add a setting for overriding the
"Reply-To" address; it could be useful for redirecting responses to an
mailbox that's actually monitored
(https://www.github.com/NZOI/nztrain/pull/272#issuecomment-2704959896).
But for now we don't bother.